+++ packages/agent/test/routes/modification/update-relation.test.ts
@@ -1,4 +1,4 @@
-import { Filter } from '@forestadmin/datasource-toolkit';
+import { Aggregation, Filter } from '@forestadmin/datasource-toolkit';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 
 import UpdateRelationRoute from '../../../src/routes/modification/update-relation';
@@ -36,7 +36,7 @@ describe('UpdateRelationRoute', () => {
     });
     const dataSource = factories.dataSource.buildWithCollections([persons, books]);
 
-    return { dataSource, services, options, router };
+    return { dataSource, services, options, router, persons, books };
   };
 
   test('should register PUT route', () => {
@@ -141,40 +141,42 @@ describe('UpdateRelationRoute', () => {
         const services = factories.forestAdminHttpDriverServices.build();
         const options = factories.forestAdminHttpDriverOptions.build();
         const router = factories.router.mockAllMethods().build();
-        const dataSource = factories.dataSource.buildWithCollections([
-          factories.collection.build({
-            name: 'book',
-            schema: factories.collectionSchema.build({
-              fields: {
-                id: factories.columnSchema.uuidPrimaryKey().build(),
-                owner: factories.oneToOneSchema.build({
-                  foreignCollection: 'owner',
-                  originKey: 'bookId',
-                  originKeyTarget: 'id',
-                }),
-              },
-            }),
-          }),
-          factories.collection.build({
-            name: 'owner',
-            schema: factories.collectionSchema.build({
-              fields: {
-                id: factories.columnSchema.uuidPrimaryKey().build(),
-                bookId: factories.columnSchema.build(),
-                book: factories.manyToOneSchema.build({
-                  foreignCollection: 'book',
-                  foreignKey: 'bookId',
-                }),
-              },
-            }),
-          }),
-        ]);
 
-        return { dataSource, services, options, router };
+        const books = factories.collection.build({
+          name: 'book',
+          schema: factories.collectionSchema.build({
+            fields: {
+              id: factories.columnSchema.uuidPrimaryKey().build(),
+              owner: factories.oneToOneSchema.build({
+                foreignCollection: 'owner',
+                originKey: 'bookId',
+                originKeyTarget: 'id',
+              }),
+            },
+          }),
+        });
+
+        const owners = factories.collection.build({
+          name: 'owner',
+          schema: factories.collectionSchema.build({
+            fields: {
+              id: factories.columnSchema.uuidPrimaryKey().build(),
+              bookId: factories.columnSchema.build(),
+              book: factories.manyToOneSchema.build({
+                foreignCollection: 'book',
+                foreignKey: 'bookId',
+              }),
+            },
+          }),
+        });
+
+        const dataSource = factories.dataSource.buildWithCollections([books, owners]);
+
+        return { dataSource, services, options, router, books, owners };
       };
 
       test('should remove the one-to-one association', async () => {
-        const { services, dataSource, options } = setupWithOneToOneRelation();
+        const { services, dataSource, options, owners } = setupWithOneToOneRelation();
 
         const update = new UpdateRelationRoute(services, options, dataSource, 'book', 'owner');
         const context = createMockContext({
@@ -186,27 +188,40 @@ describe('UpdateRelationRoute', () => {
           requestBody: {},
         });
 
+        (owners.aggregate as jest.Mock).mockResolvedValueOnce([{ value: 1 }]);
+
         await update.handleUpdateRelationRoute(context);
 
-        expect(dataSource.getCollection('owner').update).toHaveBeenCalledTimes(1);
-        expect(dataSource.getCollection('owner').update).toHaveBeenCalledWith(
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
-          new Filter({
-            conditionTree: factories.conditionTreeLeaf.build({
-              operator: 'Equal',
-              value: '00000000-0000-4000-8000-000000000000',
-              field: 'bookId',
-            }),
+        const expectedFilter = new Filter({
+          conditionTree: factories.conditionTreeLeaf.build({
+            operator: 'Equal',
+            value: '00000000-0000-4000-8000-000000000000',
+            field: 'bookId',
           }),
-          { bookId: null },
+        });
+
+        const caller = {
+          email: 'john.doe@domain.com',
+          requestId: expect.any(String),
+          timezone: 'Europe/Paris',
+        };
+
+        expect(owners.aggregate).toHaveBeenCalledWith(
+          caller,
+          expectedFilter,
+          new Aggregation({ operation: 'Count' }),
+          1,
         );
+
+        expect(owners.update).toHaveBeenCalledTimes(1);
+        expect(owners.update).toHaveBeenCalledWith(caller, expectedFilter, { bookId: null });
 
         expect(context.response.status).toEqual(HttpCode.NoContent);
         expect(services.authorization.assertCanEdit).toHaveBeenCalledWith(context, 'owner');
       });
 
       test('should change the one-to-one association', async () => {
-        const { services, dataSource, options } = setupWithOneToOneRelation();
+        const { services, dataSource, options, owners } = setupWithOneToOneRelation();
 
         const update = new UpdateRelationRoute(services, options, dataSource, 'book', 'owner');
         const context = createMockContext({
@@ -220,24 +235,114 @@ describe('UpdateRelationRoute', () => {
           },
         });
 
+        (owners.aggregate as jest.Mock).mockResolvedValueOnce([{ value: 1 }]);
+
         await update.handleUpdateRelationRoute(context);
 
-        expect(dataSource.getCollection('owner').update).toHaveBeenNthCalledWith(
+        const expectedFilter = new Filter({
+          conditionTree: factories.conditionTreeBranch.build({
+            aggregator: 'And',
+            conditions: [
+              factories.conditionTreeLeaf.build({
+                operator: 'Equal',
+                value: '00000000-0000-4000-8000-000000000000',
+                field: 'bookId',
+              }),
+              factories.conditionTreeLeaf.build({
+                operator: 'NotEqual',
+                value: '11111111-1111-4111-8111-111111111111',
+                field: 'id',
+              }),
+            ],
+          }),
+        });
+
+        const caller = {
+          email: 'john.doe@domain.com',
+          requestId: expect.any(String),
+          timezone: 'Europe/Paris',
+        };
+
+        expect(owners.aggregate).toHaveBeenCalledWith(
+          caller,
+          expectedFilter,
+          new Aggregation({ operation: 'Count' }),
           1,
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+        );
+
+        expect(owners.update).toHaveBeenNthCalledWith(1, caller, expectedFilter, { bookId: null });
+
+        expect(owners.update).toHaveBeenNthCalledWith(
+          2,
+          caller,
           new Filter({
             conditionTree: factories.conditionTreeLeaf.build({
               operator: 'Equal',
-              value: '00000000-0000-4000-8000-000000000000',
-              field: 'bookId',
+              value: '11111111-1111-4111-8111-111111111111',
+              field: 'id',
             }),
           }),
-          { bookId: null },
+          { bookId: '00000000-0000-4000-8000-000000000000' },
         );
 
-        expect(dataSource.getCollection('owner').update).toHaveBeenNthCalledWith(
-          2,
-          { email: 'john.doe@domain.com', requestId: expect.any(String), timezone: 'Europe/Paris' },
+        expect(context.response.status).toEqual(HttpCode.NoContent);
+        expect(services.authorization.assertCanEdit).toHaveBeenCalledWith(context, 'owner');
+      });
+
+      // eslint-disable-next-line max-len
+      test('should not remove the association on the new record to avoid errors with notNull constraints', async () => {
+        const { services, dataSource, options, owners } = setupWithOneToOneRelation();
+
+        const update = new UpdateRelationRoute(services, options, dataSource, 'book', 'owner');
+        const context = createMockContext({
+          state: { user: { email: 'john.doe@domain.com' } },
+          customProperties: {
+            query: { timezone: 'Europe/Paris' },
+            params: { parentId: '00000000-0000-4000-8000-000000000000' },
+          },
+          requestBody: {
+            data: { id: '11111111-1111-4111-8111-111111111111', type: 'owner' },
+          },
+        });
+
+        (owners.aggregate as jest.Mock).mockResolvedValueOnce([{ value: 0 }]);
+
+        await update.handleUpdateRelationRoute(context);
+
+        const expectedFilter = new Filter({
+          conditionTree: factories.conditionTreeBranch.build({
+            aggregator: 'And',
+            conditions: [
+              factories.conditionTreeLeaf.build({
+                operator: 'Equal',
+                value: '00000000-0000-4000-8000-000000000000',
+                field: 'bookId',
+              }),
+              factories.conditionTreeLeaf.build({
+                operator: 'NotEqual',
+                value: '11111111-1111-4111-8111-111111111111',
+                field: 'id',
+              }),
+            ],
+          }),
+        });
+
+        const caller = {
+          email: 'john.doe@domain.com',
+          requestId: expect.any(String),
+          timezone: 'Europe/Paris',
+        };
+
+        expect(owners.aggregate).toHaveBeenCalledWith(
+          caller,
+          expectedFilter,
+          new Aggregation({ operation: 'Count' }),
+          1,
+        );
+
+        expect(owners.update).toHaveBeenCalledOnce();
+        expect(owners.update).toHaveBeenCalledWith(
+          caller,
           new Filter({
             conditionTree: factories.conditionTreeLeaf.build({
               operator: 'Equal',