+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,44 @@
+## @forestadmin/agent [1.31.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.31.0...@forestadmin/agent@1.31.1) (2023-09-22)
+
+
+### Bug Fixes
+
+* create records with a hasOne relationship and a notNull constraint without error ([#826](https://github.com/ForestAdmin/agent-nodejs/issues/826)) ([32457b7](https://github.com/ForestAdmin/agent-nodejs/commit/32457b77a43612ed3164a5e8b0498e09f1bad607))
+
+# @forestadmin/agent [1.31.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.30.0...@forestadmin/agent@1.31.0) (2023-09-22)
+
+
+### Features
+
+* **widgets:** add support for the json editor widget in actions ([#830](https://github.com/ForestAdmin/agent-nodejs/issues/830)) ([cbd2661](https://github.com/ForestAdmin/agent-nodejs/commit/cbd26616af62846d54a596ce14fe972b8df73bab))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.31.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.24.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.20.0
+
+# @forestadmin/agent [1.30.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.29.2...@forestadmin/agent@1.30.0) (2023-09-21)
+
+
+### Features
+
+* dynamic search support for dropdown widget in smart action forms ([#810](https://github.com/ForestAdmin/agent-nodejs/issues/810)) ([92460a8](https://github.com/ForestAdmin/agent-nodejs/commit/92460a86fb4504047d35203440cc9a5e96af599b))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.30.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.23.0
+* **@forestadmin/forestadmin-client:** upgraded to 1.19.0
+
 ## @forestadmin/agent [1.29.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.29.1...@forestadmin/agent@1.29.2) (2023-09-20)